move delete file to common test
