+++ chain/x/oracle/common_test.go
@@ -1,8 +1,15 @@
 package oracle_test
 
 import (
+	"crypto/sha256"
+	"encoding/hex"
+	"os"
+	"path/filepath"
+
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/spf13/viper"
 	abci "github.com/tendermint/tendermint/abci/types"
+	"github.com/tendermint/tendermint/libs/cli"
 
 	bandapp "github.com/bandprotocol/bandchain/chain/app"
 	"github.com/bandprotocol/bandchain/chain/simapp"
@@ -22,4 +29,14 @@ func createTestInput() (*bandapp.BandApp, sdk.Context, me.Keeper) {
 	app := simapp.NewSimApp()
 	ctx := app.BaseApp.NewContext(false, abci.Header{})
 	return app, ctx, app.OracleKeeper
+}
+
+func deleteFile(data []byte) {
+	hash := sha256.Sum256(data)
+	filename := hex.EncodeToString(hash[:])
+	path := filepath.Join(viper.GetString(cli.HomeFlag), "files", filename)
+	err := os.Remove(path)
+	if err != nil {
+		panic(err)
+	}
 }
